UPSTREAM: x86/resctrl: Remove set_mba_sc()s control array re-initialisation
set_mba_sc() enables the 'software controller' to regulate the
bandwidth based on the byte counters. This can be managed entirely
in the parts of resctrl that move to /fs/, without any extra
support from the architecture specific code.
set_mba_sc() is called by rdt_enable_ctx() during mount and
umount. It currently resets the arch code's ctrl_val[] and mbps_val[]
arrays.

The ctrl_val[] was already reset when the domain was created,
and by reset_all_ctrls() when the filesystem was last umounted.
Doing the work in set_mba_sc() is not necessary as the values are
already at their defaults due to the creation of the domain, or were
previously reset during umount(), or are about to reset during umount().

Add a reset of the mbps_val[] in reset_all_ctrls(), allowing the
code in set_mba_sc() that reaches in to the architecture specific
structures to be removed.

Reviewed-by: Jamie Iles <quic_jiles@quicinc.com>
Tested-by: Xin Hao <xhao@linux.alibaba.com>
Reviewed-by: Shaopeng Tan <tan.shaopeng@fujitsu.com>
Tested-by: Shaopeng Tan <tan.shaopeng@fujitsu.com>
Tested-by: Cristian Marussi <cristian.marussi@arm.com>
Reviewed-by: Reinette Chatre <reinette.chatre@intel.com>
Signed-off-by: James Morse <james.morse@arm.com>