UPSTREAM: x86/resctrl: Make rdt_enable_key the arch's decision to switch
rdt_enable_key is switched when resctrl is mounted. It was also previously
used to prevent a second mount of the filesystem.

Any other architecture that wants to support resctrl has to provide
identical static keys.

Now that we have helpers for the alloc/mon keys, resctrl doesn't need
to switch this extra key, it can be done by the arch code. To allow
the static key increment/decrement functions to be used, ensure the
calls to the helpers are balanced.

Signed-off-by: James Morse <james.morse@arm.com>