DOWNSTREAM: arm64: dts/tegra: Enable gp10b and gv11b
Enable gp10b and gv11b devices.

Bug 3030537

Change-Id: Ib2b166db999347fcf1c2ecdcc97e5176d54726bf
Reviewed-on: https://git-master.nvidia.com/r/c/linux-stable/+/2376655
GVS: Gerrit_Virtual_Submit
Tested-by: Terje Bergstrom <tbergstrom@nvidia.com>
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-stable/+/2379782
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Rohit Khanna <rokhanna@nvidia.com>
Tested-by: Rohit Khanna <rokhanna@nvidia.com>

DOWNSTREAM: Add fuse clock to gp10b

Add fuse clock to gp10b. Fuse clock is required for GPU to initialize
correctly.

Bug 3030537

Change-Id: I9df85e7ad0b7c6100011c40639c99a90d7e4cd5d
Reviewed-on: https://git-master.nvidia.com/r/c/linux-stable/+/2376672
GVS: Gerrit_Virtual_Submit
Tested-by: Terje Bergstrom <tbergstrom@nvidia.com>
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-stable/+/2379783
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Rohit Khanna <rokhanna@nvidia.com>
Tested-by: Rohit Khanna <rokhanna@nvidia.com>

DOWNSTREAM: arm64: tegra: Remove unused GPU node

Commit 1c08f8aa485a ("DOWNSTREAM: arm64: tegra: Enable gp10b and gv11b")
add nodes to enable the GPU on Jetson TX2 and Jetson Xavier. Although
the GPU on Jetson TX2 is not enabled in upstream by default, the GPU on
Jetson Xavier is and so it is not necessary to populate the status
property for GV11B to enable. Furthermore, for GV11B the GPU node is
under the 'cbb@0' bus node and so the node added by the above commit
is not in the correct place.

Bug 3030537

Change-Id: Ia617587de3a2ced3062257d3e6bdb42fe9140e09
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-stable/+/2419242
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>