UPSTREAM: x86/resctrl: Allow arch to allocate memory needed in resctrl_arch_rmid_read()
Depending on the number of monitors available, Arm's MPAM may need to
allocate one prior to reading the monitor value. Allocating a contended
resource may involve sleeping.

When MPAMs bandwidth counters are over-subscribed, the count will
only increment while there is a monitor allocated. To allow a future
'resctrl_pmu' to allocate a monitor on demand, the allocate/free
calls need to be exposed to resctrl. This would supplement resctrl's
always on, free running counters, exposed to user-space, which can only
be support with MPAM if the platform has as many monitors as it does
partid*pmg. (which is unlikely).
All callers of resctrl_arch_rmid_read() read the counter on more than
one domain. If the monitor is allocated globally, there is no need to
allocate and free it for each call to resctrl_arch_rmid_read().

Add arch hooks for this allocation, which need calling before
resctrl_arch_rmid_read(). The result is passed to
resctrl_arch_rmid_read(), then freed again afterwards. The helper
can be called on any CPU, and can sleep.

Signed-off-by: James Morse <james.morse@arm.com>