iommu/arm-smmu-v3: Implement domain_alloc_user
The domain_alloc_user callback function is used to allocate iommu_domains
for user space. The input user_data is the struct iommu_user_arm_smmuv3,
which contains configurations for a stage-1 or stage-2 iommu_domain.

In nesting mode, the stage-1 configurations are included in the user_data
that allows the 2-stage STE to be done in the arm_smmu_write_strtab_ent().

Change-Id: Iafc40ba4e2c2edaab8db339e9ce5e5d3b3735661
Signed-off-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>