UPSTREAM: irqchip/gic-v3: Build PPI partitions on ACPI systems
For platforms described via DT, the gic-v3 driver allows PPIs to be
partitioned, allowing different drivers to request the PPI in each
partition. This is primarily useful on big-little systems where the
device triggering the PPI is different between partitions.

This is also useful when it describes the hardware. If a device is
integrated with a group of CPUs, a PPI interrupt it generates can
only be received on those CPUs.

The MPAM spec allows MPAM devices to trigger a PPI if the device is
integrated with a group of CPUs. This is a better choice than a
shared interrupt (e.g. SPI) as the device may not be accesible from
all CPUs.

For DT systems, PPI parititions allow the driver to treat these
devices independently.
For ACPI systems, PPI partitions are not available. Only one driver
may request a PPI. If firwmare describes multiple devices that
trigger the PPI (each for a different group of CPUs), the driver
must reference count the PPI it has already requested, update the
percpu driver argument, and handle the PPI firing on CPUs that
are mid-discovery of the corresponding device.

As this is caused by a disparity between DT and ACPI systems,
add support for PPI partitions on ACPI systems.

A Processor-Container is ACPIs way of describing a set of CPUs.
Each processor-container has a unique id. Count the processor-containers,
and set up a partition for each.

ACPI can then register a partitioned PPI by providing the intid
and the uid of the process-container that identifies the partition.

Signed-off-by: James Morse <james.morse@arm.com>