UPSTREAM: x86/resctrl: Allow the allocator to check if a closid can allocate clean rmid
MPAM's monitoring ids extend its partition ids, meaning the same id can be
used for different control groups.

This means once a closid is allocated, all its monitoring ids may still be
dirty, and held in limbo.

Add a helper to allow the closid allocator to check if a closid has dirty
rmid values. This behaviour is enabled by a kconfig option selected by
the architecture, which avoids a pointless search for x86.

Signed-off-by: James Morse <james.morse@arm.com>