iommu/tegra241-cmdqv: Add user-space use support
The CMDQV HW supports a user-space use for virtualization cases. It allows
VM to issue guest-level TLBI or ATC_INV commands directly to the queue and
executes those commands directly without a VMEXIT. The magic is done by HW
replacing the VMID field in a TLBI command and the SID field in an ATC_INV
command with the preprogrammed VMID/SID.

This is built upon the vIOMMU infrastructure by allowing VMM to allocate a
VINTF (as a vIOMMU object) and assign VCMDQs (vCMDQ objects) to it.

Each VINTF has two 64KB MMIO pages (128B per logical vCMDQ):
 - Page0 (directly accessed by guest) has all the control and status bits.
 - Page1 (trapped by VMM) has guest-owned queue memory location/size info.

VMM should trap the emulated VINTF0's page1 of the guest VM for the guest
level VCMDQ location/size info, forward that to the host kernel for vCMDQ
object allocations. Then, it should mmap the physical VINTF's page0 to the
VINTF0's page0 of the guest VM. This allows the guest OS to read and write
the guest-own VINTF's page0 for a direct control of the VCMDQ HW.

For ATC invalidation command that typically holds an SID field, it require
all passthrough devices must register their virtual SIDs to the SID_MATCH
registers and their physical SIDs to the pairing SID_REPLACE registers, so
that HW can use those as a lookup table to replace the virtual SIDs with
the correct physical SIDs.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>