PCI: Suppress kernel DMA ownership auto-claiming for bridges
IOMMU grouping on PCI necessitates that if we lack isolation on a bridge
then all of the downstream devices will be part of the same IOMMU group
as the bridge. Thus, as long as the bridge drivers don't issue any DMA
(no MMIO BAR spaces map and access) it's safe to suppress the kernel DMA
ownership auto-claiming.

The commit 5f096b14d421b "vfio: Whitelist PCI bridges" permited a class
of kernel drivers, and now that we have driver self-declaration of safety
we should rely on that, not this hack. This change may cause regression
on some platforms, since all bridges were exempted before, but now they
have to be manually audited before doing so. This is actually the desired
outcome anyway.

Suggested-by: Jason Gunthorpe <jgg@nvidia.com>
Suggested-by: Kevin Tian <kevin.tian@intel.com>
Signed-off-by: Lu Baolu <baolu.lu@linux.intel.com>