iommu: Expose group variants of dma ownership interfaces
The vfio needs to set DMA_OWNER_USER for the entire group when attaching
it to a vfio container. So expose group variants of setting/releasing dma
ownership for this purpose.

This also exposes the helper iommu_group_viable_for_user() for vfio to
report to userspace if the group is viable to user assignment, for
compatability with VFIO_GROUP_FLAGS_VIABLE.

Signed-off-by: Lu Baolu <baolu.lu@linux.intel.com>