vfio: Introduce VFIO_DEVICE_BIND_IOMMUFD
This patch introduces VFIO_DEVICE_BIND_IOMMUFD to the parked fops, as
the only contract for the user to acquire device access permission via
/dev/vfio/devices/vfioX. The success of this command indicates that
the target device has been placed in a context secure for user access
and the parked fops can be replaced by the real fops.

The presence of .{un}bind_iommufd() callbacks marks whether this command
is supported for a given vfio device type.

Open:
- The usage of replace_fops() breaks its assumption which is only used in
  ->open() instances. Needs to ensure it's harmless to use it in the
  ->ioctl() instance.
- There is also a choice to replace device->ops instead of manipulating
  filep->f_ops. This patch is going with replacing fops. Let's have more
  disccusion base on this open.

  Jason has already shared some idea via mails. Let's close it in this
  review.

Suggested-by: Jason Gunthorpe <jgg@nvidia.com>
Signed-off-by: Liu Yi L <yi.l.liu@intel.com>