Cover-letter: Add cdev for vfio_device
This series is a prerequisite for adapting vfio to the coming iommufd
model, which requires vfio to expose device-centric interface. Thus
userspace can open device fd without any group operation as current
vfio user does. But the device access via the fd should not be granted
until a secure context has been setup. This series adopts a logic of
parked fops which only provides the iommufd_bind operation. After binding
to an iommufd successfully, kernel installs a real fop to the device
fd to grant device access.

But we have an open on the parked fops. Maybe replacing the vfio_device_ops
works as well. Expects to close this open in the version.

This series was based on the dma_owner patches from Baolu and rebased on
Alex's vfio-v5.16-rc1 (https://github.com/awilliam/linux-vfio/tree/vfio-v5.16-rc1)
A complete version can be found in below branch (also includes iommufd
patches).

https://github.com/luxis1999/iommufd/tree/iommufd-vfio-v5.16-rc1

Signed-off-by: Liu Yi L <yi.l.liu@intel.com>