iommufd: Data structure to provide IOVA to PFN mapping
This is an implementation of a similar data structure in
drivers/vfio/vfio_iommu_type1.c. It is intended to be compatible with
those semantics while also offering a more modern design and a few
additional features.

At the core is an object called iopt_pages which stores a raw list of
PFNs, obtained from userspace. iopt_pages can hold the page list inside an
iommu_domain, or inside an xarray if no iommu_domain is available.

Each iopt_pages can be refered by multiple top level io_pagetables
allowing sharing of both the memory holding the PFNs and the pin counts
themselves. This allows userspace to operate in a model where it has only
one pin per memory area and can construct many different IOMMU
configurations from it.

The io_pagetable's have an interval tree of iopt_area's which describe the
populated IOVA ranges and link to the iopt_pages that provide the PFNs.

iommu_domains can be connected to an io_pagetable, and will be populated
with PFNs either by copying from other iommu_domains, the xarray, or
freshly pinning pages.

SW users can access a struct page array of PFNs under the IOVA's via a
pinning mechanism. This mechanism can also source pages from the
iommu_domain, xarray or freshly pinning.

Compared to vfio_iommu_type1.c there are several limitations:
 - Only the new mode is supported where unmap cannot punch holes or
   split areas
 - Resource limits rely on memory cgroups to bound what userspace can do
   instead of the module parameter dma_entry_limit
 - VFIO P2P is not implemented
 - Fast RW is not implemented
 - Indefinite suspend of SW access (VFIO_DMA_MAP_FLAG_VADDR)
 - FIXME: Need to audit more carefully and look for differences

This design suggests several possible imporvements in the IOMMU API that
would greatly help peformance, particularly a way to read PFNs from an
iommu_domain that is not a function indirection per 4k PFN, and a way for
a driver to implement the domain copy.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>