iommufd: IOCTLs for the io_pagetable
Connect the IO Address Space Pagetable to its IOCTL interface. This
exposes most of the functionality in the io_pagetable to userspace.

This is inteded to be the core of the generic interface that IOMMUFD will
provide. Every IOMMU driver should be able to implement an iommu_domain
that is compatible with this generic mechanism.

It is also designed to be easy to use for simple non
virtual-machine-monitor users, like DPDK. The kernel provides universal
support for all IOMMUs no (PPC special path), an IOVA allocator, and
allows any number or configuration of devices to be connected to the IOAS,
simplifying these applications.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>