vfio: Add cdev for vfio_device
This allows user to directly open a vfio device w/o using the legacy
container/group interface, as a prerequisite for iommufd. However, the
device fd opened via this new interface don't have the capability to
access the device as the fops open() doesn't set a valid value to the
private_data of the file. Next patch will add one command (BIND_IOMMUFD)
to the ioctl of this fd. Once it succeeds, a valid private data will be
set to unblock the user access.

With this patch, devices registered to vfio core have both group and device
interface created.

- group interface : /dev/vfio/$groupID
- device interface: /dev/vfio/devices/vfioX  (X is the minor number)

Given a vfio device the user can identify the matching vfioX by checking
the sysfs path of the device. Take PCI device (0000:6a:01.0) for example,
/sys/bus/pci/devices/0000\:6a\:01.0/vfio-device/vfio0/dev contains the
major:minor of the matching vfioX.

Userspace then opens the /dev/vfio/devices/vfioX and checks with fstat
that the major:minor matches.

The cdev logic in this patch:
*) __vfio_register_dev() path ends up doing cdev_device_add() for each
   vfio_device;
*) __vfio_unregister_dev() path does cdev_device_del();

Signed-off-by: Liu Yi L <yi.l.liu@intel.com>