vfio: Add VFIO_DEVICE_BIND_IOMMUFD support
This patch introduces VFIO_DEVICE_BIND_IOMMUFD, as the only contract for
the user to acquire device access permission via /dev/vfio/devices/vfioX.
The success of this command indicates that the target device has been
placed in a context secure for user access and the user access via this fd
is granted.

The presence of .{un}bind_iommufd() callbacks marks whether this command
is supported for a given vfio device type.

Change-Id: Ic4e0a8d81558e2e07ba990592ec9f0157eb063da
Signed-off-by: Yi Liu <yi.l.liu@intel.com>
[nicolinc: we need to fix device pointer using cdev]
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>