vfio: Add iommufd VFIO compat support to group_fd
IOMMUFD has VFIO compatibility so user space can open /dev/iommu
to reuse existing VFIO IOMMU ioctls. However, the group_fd used
for VFIO_GROUP_SET/UNSET_CONTAINER is still a VFIO thing and is
not supported using new IOMMUFD ioctls, which go for an approach
of attaching and dettaching a VFIO device directly instead of a
VFIO group.

In order to fill this gap, IOMMUFD provides a pair of helpers to
VFIO to call, so as to bind/unbind all devices in the VFIO group
to/from the iommufd and to attach/detach them to/from the ioas.

So this patch simply calls those two helpers in cases where user
space uses VFIO ioctls on top of /dev/iommu and wants to connect
group_fd to that.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Signed-off-by: Yi Liu <yi.l.liu@intel.com>