iommufd: Add vfio_group_set/unset_iommufd helpers
IOMMUFD has VFIO compatibility so user space can open /dev/iommu
to reuse existing VFIO IOMMU ioctls. However, the group_fd used
for VFIO_GROUP_SET/UNSET_CONTAINER is still a VFIO thing and is
not supported using new IOMMUFD ioctls, which go for an approach
of attaching and dettaching a VFIO device directly instead of a
VFIO group.

In order to fill this gap, this patch adds a pair of helpers for
VFIO to call, so as to bind/unbind all devices in the VFIO group
to/from the iommufd and to attach/detach them to/from the ioaspt.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>