vfio/iommu_type1: Reorganize vfio_iommu_type1_attach_group()
The vfio_iommu_type1_attach_group() allocates and attaches a
new domain, prior to matching an existing one. This ended up
with having a back-to-back detach and attach routine:
	new = alloc();
	new->domain = iommu_domain_alloc();
	iommu_attach_group(new, ...);
	list_for_each_entry (old, existing_list, ..) {
		if (!match(old, new->domain))
			continue;
		detach(new->domain);
		if (!attach(old))
			// reuse old one
		attach(new->domain);
	}

This can be avoided by using a test domain at the beginning:
	test = iommu_domain_alloc();
	list_for_each_entry (old, existing_list, ..) {
		if (!match(old, new->domain))
			continue;
		if (attach(old))
			// reuse old one
	}
	// otherwise
	new = alloc();
	new->domain = test;

So, this patch reorganizes the vfio_iommu_type1_attach_group()
by moving most of the new domain routine into a new function,
following the idea above.

It also adds an msi_cookie flag in the struct vfio_domain, to
protect iommu_get_msi_cookie() call in the main routine, since
now existing domain would run into that too.

This is also a prepartory change for an upcoming series that
will add an iommu helper function to check whether the domain
can be attached before doing actual iommu_attach_group().

Change-Id: I48339c0e8caaa7e3902c0e26ad3dfd5094ee8915
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>