vfio/iommu_type1: Reorganize vfio_iommu_type1_attach_group()
There is a reversed domain attaching routine in this function,
where it allocates and attaches a new domain prior to matching
an existing domain, which resulted in a detach-attach-reattach
situation. This can be avoide by trying a testing iommu_domain
in the match routine.

So, this patch reorganizes the vfio_iommu_type1_attach_group()
in the following steps:
0) Get reserved regions and MSI info of pre-attached iommu_group
1) Allocate a new testing iommu_domain: new_domain
2) Match an existing domain by verifying it with new_domain and
   ensuring it can be attached with an MSI cookie.
3) If no match, allocate and attach group to a new vfio_domain
4) Conflict checks on post-attached iommu_group
5) Run additional validation in case of a new vfio_domain
6) Finalize everything (previous "done:" routine)

This is also a prepartory change for a following cleanup patch
that will use iommu_can_attach_group() helper function.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>