vfio/iommu_type1: Reorganize vfio_iommu_type1_attach_group()
The vfio_iommu_type1_attach_group() allocates and attaches a
new domain, prior to matching an existing one. This ended up
with having a back-to-back detach and attach routine:
	new = alloc();
	new->domain = iommu_domain_alloc();
	iommu_attach_group(new, ...);
	list_for_each_entry (old, existing_list, ..) {
		if (!match(old, new->domain))
			continue;
		detach(new->domain);
		if (!attach(old))
			// reuse old one
		attach(new->domain);
	}

This can be avoided by using a test domain at the beginning:
	test = iommu_domain_alloc();
	list_for_each_entry (old, existing_list, ..) {
		if (!match(old, new->domain))
			continue;
		if (attach(old))
			// reuse old one
	}
	// otherwise
	new = alloc();
	new->domain = test;

So, this patch reorganizes the vfio_iommu_type1_attach_group()
by moving most of the new domain routine into a new function,
following the idea above.

The new program flow will be:
1) Get reserved regions of pre-attached iommu_group
2) Sub-function vfio_iommu_alloc_and_attach_domain()
  2-1) Allocate a new testing iommu_domain: new_domain
  2-2) Match an existing domain by verifying it with new_domain
       and ensuring it can be attached
  2-3) If no match, allocate and attach a new vfio_domain
  2-4) Finalize new domain: fgsp, replay, msi, and etc.
4) Region conflict checks on post-attached iommu_group
5) Finalize everything (previous "done:" routine)

This is also a prepartory change for an upcoming series that
will add an iommu helper function to check whether the domain
can be attached before doing actual iommu_attach_group().

Change-Id: I48339c0e8caaa7e3902c0e26ad3dfd5094ee8915
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>