vfio/iommu_type1: Reorganize vfio_iommu_type1_attach_group()
There is a reversed domain attaching routine in this function,
where it allocates and attaches a new domain prior to matching
an existing domain, which resulted in a detach-attach-reattach
situation. This can be avoide by trying a testing iommu_domain
in the match routine.

So, this patch reorganizes the vfio_iommu_type1_attach_group()
using the following routine:
1) Get reserved regions and MSI info of pre-attached iommu_group
2) Sub-function vfio_iommu_alloc_and_attach_domain()
  2-1) Allocate a new testing iommu_domain: new_domain
  2-2) Match an existing domain by verifying it with new_domain
       and ensuring it can be attached with an MSI cookie
  2-3) If no match, allocate and attach a new vfio_domain
4) Region conflict checks on post-attached iommu_group
5) Additional validation with vfio_iommu_finalize_new_domain()
6) Finalize everything (previous "done:" routine)

This is also a prepartory change for an upcoming series that
will add an iommu helper function to check whether the domain
can be attached before doing actual iommu_attach_group().

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>