UPSTREAM: fs/resctrl: Split resctrl_arch_mon_ctx_alloc() to have a no_wait version
resctrl_arch_mon_ctx_alloc() is called by resctrl to allocate an MPAM
monitor, so that it can go on to read the counters. This call needs
to be able to sleep until a monitor becomes available. This is only
used to read single values, so monitors are not allocated long term.

The resctrl_pmu needs to allocate an MPAM monitor which it will
hold onto for a longer period. The PMU event creation shouldn't block
until a monitor becomes available as this could deadlock if two
events are createde, but only one monitor is available.

Change the architecture specific hook to be a no_wait version, and
add the retry loop as a weak symbol in resctrl.

Signed-off-by: James Morse <james.morse@arm.com>